👌 [RUMF-530] split logs API and init code and tests
